requirements.txt
altgraph==0.17.4
certifi==2024.2.2        
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0        
idna==3.7
kiwisolver==1.4.5        
matplotlib==3.8.3        
numpy==1.25.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.2.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyVISA==1.13.0
pywin32-ctypes==0.2.2
requests==2.31.0
RsInstrument==1.60.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.1